issue 1823131: Support zero copy on the dst entry tcp class
For zerocopy sends, modify the dst tcp class fast send logic to
use the ring API to get lkey for user buffers.

The user buffers are assumed to originate from huge pages of certain
size, and hence we maintain one lkey per page and in run time mask
a user buffer address to the start of page to retrieve the lkey.

Force usage of HW LSO for zerocopy sends, as this is the only
mode of the eth mlx5 qp class that supports multiple SGEs.

** it's possible for the ZC payload to span only one packet,
** since the HW's default for this corner case is to avoid
** the LSO directive in the WQE, you must run once this command:
** BF  --> mcra /dev/mst/mt41682_pciconf0 0x31500.3:1 0
** CX5 --> mcra /dev/mst/mt4121_pciconf0 0x31500.3:1 0

Signed-off-by: Or Gerlitz <ogerlitz@mellanox.com>