issue 1823131: Support zerocopy tx segments
Add flags to denote zerocopy sends in multiple levels: vma, lwip tcp and
lwip tcp segment. For ZC segments we don't change the pbuf payload pointer
while prepareing for HW segmentation (TSO).

A buffer for network (tcp/l3/l2) headers is added on the segment
struct since we can't do anymore the tricks with pushing/pulling
p->payload as done for non-zc segment pbufs.

Currenlty we don't support segment splitting (small window or rexmit flow).
The new logic of tcp_rexmit_segment was adjusted for ZC segments.

Segment shrinking (ack flow) is supported but we either release or keep
whole pbufs, not doing the partial shrink which involves a copy.

One SGE entry is resevered to contain the network (l2/l3/tcp) headers
for ZC segments. In that case, just before lwip calls back to vma's
ip_output to send the packet, a dedicated pbuf is allocated, the headers
are copied there, and then we prepend it to the segment's pbuf chain.

For non TSO config, allow to use up to four and not two SG entries,
this should be probably enough to fill MSS sized segments when
running ZC with tso off.

Signed-off-by: Or Gerlitz <ogerlitz@mellanox.com>